Cleaning Network modules step 1
[both the client and sever:]
1. Packet.cs -> common/Packet.cs
2. ThreadManager.cs -> ClientThreadManager.cs & ServerThreadManager.cs
[client side:]
3. Add a inputfield for username input
4. Client.cs -> ClientInstance.cs. Change the class name accordingly.
5. Create ClientGameConstants.cs for GameLogic related constants.
6. namespace "Network.Client" "Network"
