+++ Assets/Project/Scenes/NetworkingDemo-Server.unity
@@ -1220,8 +1220,27 @@ PrefabInstance:
       propertyPath: spawnLocation.y
       value: 1
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 4426444533257792514, guid: d1aca9afa8776ee4e8f39a34197b5f43, type: 3}
   m_SourcePrefab: {fileID: 100100000, guid: d1aca9afa8776ee4e8f39a34197b5f43, type: 3}
+--- !u!1 &4426444534065418798 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 4426444533257792515, guid: d1aca9afa8776ee4e8f39a34197b5f43,
+    type: 3}
+  m_PrefabInstance: {fileID: 4426444534065418797}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &4426444534065418799
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4426444534065418798}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 407d6aca0d4374365af500096d826e11, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1001 &8077959778646417329
 PrefabInstance:
   m_ObjectHideFlags: 0
