+++ Assets/Project/Scenes/NetworkingDemo.unity
@@ -1111,7 +1111,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1164642867}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 395dc8cc036bd4144a33fda9ec4abeb9, type: 3}
+  m_Script: {fileID: 11500000, guid: 679a26f33dc4940e19bac5a9fcf121c9, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!114 &1164642870
@@ -1842,6 +1842,16 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 884728583}
     m_Modifications:
+    - target: {fileID: 2222572675978920758, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: -180
+      objectReference: {fileID: 0}
+    - target: {fileID: 2222572676197230313, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: -60
+      objectReference: {fileID: 0}
     - target: {fileID: 2222572676540410501, guid: 4a6759160d9117b42b5ede9abcf370ab,
         type: 3}
       propertyPath: m_LocalPosition.x
@@ -1952,5 +1962,59 @@ PrefabInstance:
       propertyPath: m_Name
       value: ClientPCNetworkMenu
       objectReference: {fileID: 0}
+    - target: {fileID: 2222572676540410506, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: portField
+      value: 
+      objectReference: {fileID: 2222572675519305461}
+    - target: {fileID: 2222572676540410506, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: usernameField
+      value: 
+      objectReference: {fileID: 2222572675519305460}
+    - target: {fileID: 2222572676971085855, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: -120
+      objectReference: {fileID: 0}
+    - target: {fileID: 4229159534373902544, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: m_AnchoredPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4229159534373902544, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4229159534373902544, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: m_SizeDelta.x
+      value: 600
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 4a6759160d9117b42b5ede9abcf370ab, type: 3}
+--- !u!114 &2222572675519305460 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 3322108030054728684, guid: 4a6759160d9117b42b5ede9abcf370ab,
+    type: 3}
+  m_PrefabInstance: {fileID: 2222572675519305459}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d199490a83bb2b844b9695cbf13b01ef, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &2222572675519305461 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 571180525612315679, guid: 4a6759160d9117b42b5ede9abcf370ab,
+    type: 3}
+  m_PrefabInstance: {fileID: 2222572675519305459}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d199490a83bb2b844b9695cbf13b01ef, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
