+++ Assets/Project/Scenes/OperatingRoom-Client.unity
@@ -1152,6 +1152,24 @@ Transform:
   m_Father: {fileID: 1158019259}
   m_RootOrder: 9
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1 &248471474 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 4531919474393089605, guid: 1893d7955d41e854186a9bec83d04194,
+    type: 3}
+  m_PrefabInstance: {fileID: 633248415}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &248471475
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 248471474}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 679a26f33dc4940e19bac5a9fcf121c9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &285456069
 GameObject:
   m_ObjectHideFlags: 0
@@ -2024,7 +2042,8 @@ PrefabInstance:
       propertyPath: m_Name
       value: ClientManager
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 4531919474393089603, guid: 1893d7955d41e854186a9bec83d04194, type: 3}
   m_SourcePrefab: {fileID: 100100000, guid: 1893d7955d41e854186a9bec83d04194, type: 3}
 --- !u!4 &633248416 stripped
 Transform:
@@ -6135,6 +6154,16 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 1338650458}
     m_Modifications:
+    - target: {fileID: 2222572675978920758, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: -180
+      objectReference: {fileID: 0}
+    - target: {fileID: 2222572675978920758, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: m_AnchoredPosition.x
+      value: 220
+      objectReference: {fileID: 0}
     - target: {fileID: 2222572676092330630, guid: 4a6759160d9117b42b5ede9abcf370ab,
         type: 3}
       propertyPath: m_SizeDelta.x
@@ -6144,6 +6173,11 @@ PrefabInstance:
         type: 3}
       propertyPath: m_AnchoredPosition.x
       value: -75
+      objectReference: {fileID: 0}
+    - target: {fileID: 2222572676197230313, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: -60
       objectReference: {fileID: 0}
     - target: {fileID: 2222572676540410501, guid: 4a6759160d9117b42b5ede9abcf370ab,
         type: 3}
@@ -6260,6 +6294,36 @@ PrefabInstance:
       propertyPath: portField
       value: 
       objectReference: {fileID: 906700831}
+    - target: {fileID: 2222572676540410506, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: usernameField
+      value: 
+      objectReference: {fileID: 1725569142}
+    - target: {fileID: 2222572676971085855, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: -120
+      objectReference: {fileID: 0}
+    - target: {fileID: 3548241530352934880, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: m_FontData.m_Alignment
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 4229159534373902544, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: m_SizeDelta.x
+      value: 600
+      objectReference: {fileID: 0}
+    - target: {fileID: 4229159534373902544, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: m_AnchoredPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4229159534373902544, guid: 4a6759160d9117b42b5ede9abcf370ab,
+        type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: 0
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 4a6759160d9117b42b5ede9abcf370ab, type: 3}
 --- !u!224 &1578539650 stripped
@@ -7352,6 +7416,18 @@ Transform:
   m_Father: {fileID: 505617245}
   m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1725569142 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 3322108030054728684, guid: 4a6759160d9117b42b5ede9abcf370ab,
+    type: 3}
+  m_PrefabInstance: {fileID: 1578539649}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d199490a83bb2b844b9695cbf13b01ef, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1733629393
 GameObject:
   m_ObjectHideFlags: 0
