+++ Assets/Project/Scenes/OperatingRoom-Server.unity
@@ -38,7 +38,7 @@ RenderSettings:
   m_ReflectionIntensity: 1
   m_CustomReflection: {fileID: 0}
   m_Sun: {fileID: 0}
-  m_IndirectSpecularColor: {r: 0.12809429, g: 0.15871584, b: 0.21006814, a: 1}
+  m_IndirectSpecularColor: {r: 0.10940202, g: 0.13990085, b: 0.1904511, a: 1}
   m_UseRadianceAmbientProbe: 0
 --- !u!157 &3
 LightmapSettings:
@@ -1335,7 +1335,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 7
+  m_RootOrder: 6
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &432115626
 PrefabInstance:
@@ -3078,8 +3078,27 @@ PrefabInstance:
       propertyPath: spawnRotation.y
       value: -90
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 4426444533257792514, guid: d1aca9afa8776ee4e8f39a34197b5f43, type: 3}
   m_SourcePrefab: {fileID: 100100000, guid: d1aca9afa8776ee4e8f39a34197b5f43, type: 3}
+--- !u!1 &921120281 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 4426444533257792515, guid: d1aca9afa8776ee4e8f39a34197b5f43,
+    type: 3}
+  m_PrefabInstance: {fileID: 921120280}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &921120282
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 921120281}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 407d6aca0d4374365af500096d826e11, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &922421530
 GameObject:
   m_ObjectHideFlags: 0
@@ -5986,7 +6005,7 @@ Transform:
   - {fileID: 1797326438}
   - {fileID: 252570640}
   m_Father: {fileID: 0}
-  m_RootOrder: 6
+  m_RootOrder: 8
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &1643411618
 MonoBehaviour:
@@ -7067,7 +7086,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 8
+  m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1780108788
 GameObject:
