+++ Assets/Project/Scripts/GameLogic/Client.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 74ec3defd099c44b09651ccc1abc9d1c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
