+++ Assets/Project/Scripts/GameLogic/Client/ClientGameConstants.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace CEMSIM
+{
+    namespace GameLogic
+    {
+        public class ClientGameConstants
+        {
+            public const string CLIENT_DEFAULT_USERNAME = "player";
+        }
+    }
+}
