+++ Assets/Project/Scripts/GameLogic/Client/ClientGameConstants.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 03174856f09cba640bea41cd3b36626b
+guid: d2ffb9ddcf0bc42b388c2b8ca925440c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
