+++ Assets/Project/Scripts/NetworkingScripts/Client/ClientInstance.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 588a9c18856984b618845dd3bb1bf404
+guid: 08e4de13b43af4faaabbde165d89e39b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
