+++ Assets/Project/Scripts/NetworkingScripts/Client/ClientNetworkConstants.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace CEMSIM
+{
+    namespace Network
+    {
+        public class ClientNetworkConstants
+        {
+            public const string SERVER_IP = "orsimserver.ddns.net";
+            public const int SERVER_PORT = 54321;
+            public const int DATA_BUFFER_SIZE = 4096; //4MB for both Tx & Rx
+        }
+    }
+}
