--- Assets/Project/Scripts/NetworkingScripts/Client/Constants.cs
@@ -1,16 +1,0 @@
-﻿using System;
-
-namespace CEMSIM
-{
-    namespace Network
-    {
-        namespace Client
-        {
-            public class Constants
-            {
-                public const string SERVER_IP = "orsimserver.ddns.net";
-                public const int SERVER_PORT = 54321;
-            }
-        }
-    }
-}
