+++ Assets/Project/Scripts/NetworkingScripts/Common/Packet.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 86a078e357f274244b2fb53d339d4156
+guid: 3f6b2512e9a5c496eb82a12d6463e6ff
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
